Move support functions to environment_support.py
