+++ gym-drill/gym_drill/envs/Target.py
@@ -1,16 +1,19 @@
 from gym_drill.envs.Coordinate import Coordinate
 
+TARGET_REWARD = 100
+
 class TargetBall():
     def __init__(self, x, y, rad):
         self.center = Coordinate(x,y)
         self.radius = rad
         
         self.reached = False # Indicate that the target has been reached.
-        self.is_hit = False
+        self.reward = 100
 
     # Print the target to console. Can be nice for debugging purposes
     def __str__(self):
         return "Center:" + str(self.center) +". Radii:" + str(self.radius)
+
